Add job test, model, controller, routes, database migration, model factory and test utilities
This is an example of a basic Laravel workflow to be noted by the team.
In this commit, added Laravel debugbar (see composer.json) for debugging purposes, must set APP_DEBUG = true to see it on the bottom of the screen.
Can you faker to generate random data for testing purposes.